+++ src/elements/db/ElementQuery.php
@@ -1893,7 +1893,17 @@ class ElementQuery extends Query implements ElementQueryInterface
      */
     public function exists($db = null): bool
     {
-        return $this->getCachedResult() !== null || parent::exists($db);
+        $cachedResult = $this->getCachedResult();
+        if ($cachedResult !== null) {
+            return !empty($cachedResult);
+        }
+        try {
+            return $this->prepareSubquery()
+                ->select('elements.id')
+                ->exists($db);
+        } catch (QueryAbortedException) {
+            return false;
+        }
     }
 
     /**